python/requirements.txt
fastapi>=0.115.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-multipart>=0.0.12
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
